push

+++ Facebook-Frontend/src/app/shared/components/post/post-header/post-header.component.ts
@@ -1,11 +1,21 @@
-import { ChangeDetectionStrategy, Component, ElementRef, EventEmitter, HostListener, inject, input, Output } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  ElementRef,
+  EventEmitter,
+  HostListener,
+  inject,
+  input,
+  Output,
+} from '@angular/core';
 import { PostService } from 'src/app/core/services/post.service';
+import { getTimeAgoForConversation } from 'src/app/shared/utils/date-format-utils';
 
 @Component({
   selector: 'app-post-header',
   templateUrl: './post-header.component.html',
   styleUrl: './post-header.component.css',
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class PostHeaderComponent {
   postId = input.required<number>();
@@ -18,32 +28,32 @@ export class PostHeaderComponent {
   hiddenIcon = input.required<boolean>();
   options = input.required<boolean>();
   isShowPostForm: boolean = false;
-  isOptionsOpen : boolean = false;
+  isOptionsOpen: boolean = false;
   postService = inject(PostService);
   privacyIcons = [
     {
       mode: 'public',
-      icon: 'fa-solid fa-earth-americas'
+      icon: 'fa-solid fa-earth-americas',
     },
     {
       mode: 'friends',
-      icon: 'fa-solid fa-user-friends'
+      icon: 'fa-solid fa-user-friends',
     },
     {
       mode: 'only me',
-      icon: 'fa-solid fa-lock'
-    }
-  ]
+      icon: 'fa-solid fa-lock',
+    },
+  ];
   privacy = input.required<string>();
   getPrivacyIcon(privacy: string) {
-    return this.privacyIcons.find(icon => icon.mode === privacy)?.icon;
+    return this.privacyIcons.find((icon) => icon.mode === privacy)?.icon;
   }
-  getConvertedPrivacy(privacy: string){
-    if(privacy === 'friends'){
+  getConvertedPrivacy(privacy: string) {
+    if (privacy === 'friends') {
       return 'Bạn bè';
-    }else if(privacy === 'public'){
+    } else if (privacy === 'public') {
       return 'Công khai';
-    }else {
+    } else {
       return 'Chỉ mình tôi';
     }
   }
@@ -58,9 +68,16 @@ export class PostHeaderComponent {
   // Nghe sự kiện click ra ngoài để đóng dropdown
   @HostListener('document:click', ['$event'])
   onClickOutside(event: MouseEvent) {
-    const optionButton = document.getElementById('option-button'+this.postId());
-    const options = document.getElementById('options'+this.postId());
-    if (optionButton && options && !options.contains(event.target as Node) && !optionButton.contains(event.target as Node)) {
+    const optionButton = document.getElementById(
+      'option-button' + this.postId()
+    );
+    const options = document.getElementById('options' + this.postId());
+    if (
+      optionButton &&
+      options &&
+      !options.contains(event.target as Node) &&
+      !optionButton.contains(event.target as Node)
+    ) {
       this.closeOptions();
     }
   }
@@ -68,17 +85,21 @@ export class PostHeaderComponent {
   onShowPostForm() {
     this.isShowPostForm = true;
   }
-  onHidePostForm(){
+  onHidePostForm() {
     this.isShowPostForm = false;
   }
 
-  
   onDelete() {
-    console.log('Delete post:', this.postId());
-    this.postService.deletePost(this.postId()).subscribe({
-      next: (response: any) => {
-        window.location.reload();
-      },
-    });
+    const isConfirmed = window.confirm(
+      'Are you sure you want to delete this post?'
+    );
+    if (isConfirmed) {
+      this.postService.deletePost(this.postId()).subscribe({
+        next: (response: any) => {
+          alert(response.message);
+          window.location.reload();
+        },
+      });
+    }
   }
 }
